requirements.txt
requests>=2.31.0
yt-dlp>=2024.01.01
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
faster-whisper>=1.0.3
google-generativeai>=0.3.0
python-dotenv>=1.0.0
